refactor: factorise talk and tracks schema usage
