+++ app/features/event-participation/event-page/services/event-page.server.ts
@@ -2,10 +2,7 @@ import { db } from 'prisma/db.server.ts';
 import { z } from 'zod';
 import { SurveyConfig } from '~/features/event-management/settings/models/survey-config.ts';
 import { EventNotFoundError } from '~/shared/errors.server.ts';
-import {
-  TracksMandatorySchema,
-  TracksOptionalSchema,
-} from '../../cfp-submission/services/talk-submission.schema.server.ts';
+import { TracksMandatorySchema, TracksOptionalSchema } from '~/shared/types/speaker-talk.types.ts';
 
 export class EventPage {
   constructor(private slug: string) {}
