+++ app/features/speaker/talk-library/services/speaker-talk.server.ts
@@ -1,7 +1,7 @@
 import { db } from 'prisma/db.server.ts';
 import { TalkNotFoundError } from '~/shared/errors.server.ts';
 import type { Languages } from '~/shared/types/proposals.types.ts';
-import type { TalkSaveData } from './talks-library.schema.server.ts';
+import type { TalkSaveData } from '~/shared/types/speaker-talk.types.ts';
 
 export class SpeakerTalk {
   constructor(
